intelligence/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
pandas==2.2.0
openpyxl==3.1.2
pyyaml==6.0.1
python-multipart==0.0.9
numpy==1.26.4
pymupdf==1.23.8
